back/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
pymongo==4.7.2
python-dotenv==1.0.1
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
PyJWT==2.8.0
bcrypt==4.1.3
gunicorn==22.0.0
